requirements.txt
langchain_anthropic
Streamlit
Pandas
Pypdf
kor